fix(coding-agent): isolate startup cwd helper
